+++ controller/thymis_controller/models/__init__.py
@@ -1,5 +1,6 @@
 from .agent_connection import *
 from .artifacts import *
+from .auth import *
 from .device import *
 from .device_metric import *
 from .fleet_alert import *
